% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marascuilo.R
\name{autoplot.marascuilo}
\alias{autoplot.marascuilo}
\title{Autoplot method for marascuilo objects}
\usage{
\method{autoplot}{marascuilo}(object, ...)
}
\arguments{
\item{object}{An object of class \code{marascuilo}.}

\item{...}{Additional arguments.}
}
\value{
A ggplot object visualizing the pairwise comparisons.
}
\description{
Autoplot method for marascuilo objects
}
