% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marascuilo.R
\name{plot.marascuilo}
\alias{plot.marascuilo}
\title{Plot method for marascuilo objects}
\usage{
\method{plot}{marascuilo}(x, ...)
}
\arguments{
\item{x}{An object of class \code{marascuilo}.}

\item{...}{Additional arguments passed to plot.}
}
\value{
A ggplot object visualizing the pairwise comparisons.
}
\description{
Plot method for marascuilo objects
}
