% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marascuilo.R
\name{plot_marascuilo}
\alias{plot_marascuilo}
\title{Marascuilo Test Plot}
\usage{
plot_marascuilo(resultados)
}
\arguments{
\item{resultados}{A data.frame generated by the marascuilo_test() function.}
}
\value{
A ggplot object visualizing the pairwise comparisons.
}
\description{
Marascuilo Test Plot
}
\examples{
df_ejemplo <- data.frame(
  Ciudad = c(rep("Cordoba", 150), rep("Villa Maria", 75), rep("Rio IV", 75)),
  Opinion = c(
    rep("De acuerdo", 115), rep("En desacuerdo", 35),
    rep("De acuerdo", 53), rep("En desacuerdo", 22),
    rep("De acuerdo", 40), rep("En desacuerdo", 35)
  )
)
res <- marascuilo_test(df_ejemplo, "Ciudad", "Opinion", "De acuerdo")
plot_marascuilo(res)
}
