% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marascuilo.R
\name{print.marascuilo}
\alias{print.marascuilo}
\title{Print method for marascuilo objects}
\usage{
\method{print}{marascuilo}(x, ...)
}
\arguments{
\item{x}{An object of class \code{marascuilo}.}

\item{...}{Additional arguments passed to print.}
}
\description{
Print method for marascuilo objects
}
