% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marascuilo.R
\name{marascuilo_test}
\alias{marascuilo_test}
\title{Test de Marascuilo}
\usage{
marascuilo_test(df, col_grupo, col_respuesta, valor_exito, alpha = 0.05)
}
\arguments{
\item{df}{Un data.frame con los datos crudos.}

\item{col_grupo}{Nombre de la columna que contiene las categorías/grupos.}

\item{col_respuesta}{Nombre de la columna con las respuestas (ej. de acuerdo/en desacuerdo).}

\item{valor_exito}{El valor específico en col_respuesta del que se calculará la proporción.}

\item{alpha}{Nivel de significación (por defecto 0.05).}
}
\value{
Un data.frame con los resultados de las comparaciones.
}
\description{
Test de Marascuilo
}
