% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marascuilo.R
\name{plot_marascuilo}
\alias{plot_marascuilo}
\title{Gráfico del Test de Marascuilo}
\usage{
plot_marascuilo(resultados)
}
\arguments{
\item{resultados}{Un data.frame generado por la función marascuilo_test().}
}
\value{
Un objeto ggplot con la visualización de las comparaciones.
}
\description{
Gráfico del Test de Marascuilo
}
